refactor(api): modularize and optimize API handler logic
Moved endpoint handler logic into dedicated modular functions under the `apis` package to improve readability and reusability. This change centralizes the route-specific handling while reducing boilerplate in `app.py`.

- Introduced separate handlers for `/producao`, `/processamento`, `/comercializacao`, `/importacao`, and `/exportacao`.
- Extracted utility functions and data validation into the `utils.py` module.
- Updated API routes in `app.py` to use the new modular handlers, resulting in cleaner and more maintainable code.
